+++ src/components/HelpPopover/index.js
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import OverlayWrapper from '../OverlayWrapper';
 import Popover from '../Popover';
-import Icon from 'react-oui-icons';
+import ButtonIcon from '../ButtonIcon';
 
 const HelpPopover = ({
   behavior,
@@ -10,6 +10,7 @@ const HelpPopover = ({
   horizontalAttachment,
   iconSize,
   isConstrainedToScreen,
+  popoverId,
   popoverTitle,
   verticalAttachment,
   testSection,
@@ -19,21 +20,21 @@ const HelpPopover = ({
     horizontalAttachment={ horizontalAttachment }
     isConstrainedToScreen={ isConstrainedToScreen }
     overlay={
-      <Popover title={ popoverTitle }>
+      <Popover title={ popoverTitle } popoverId={ popoverId }>
         { children }
       </Popover>
     }
     testSection={ testSection }
     verticalAttachment={ verticalAttachment }>
-    <span
+    <ButtonIcon
+      ariaDescribedBy={ popoverId }
       className="cursor--pointer push-half--left"
-      style={{'marginBottom': '1px'}}
-      data-test-section="help-popover-icon">
-      <Icon
-        name="help"
-        size={ iconSize }
-      />
-    </span>
+      style="plain"
+      data-test-section="help-popover-icon"
+      iconName="help"
+      iconSize={ iconSize }
+      title="Help"
+    />
   </OverlayWrapper>
 );
 
@@ -48,7 +49,9 @@ HelpPopover.propTypes = {
   iconSize: PropTypes.oneOf(['small', 'medium', 'large']),
   /** Attach `overlay` to an edge of the screen if it is going to move off */
   isConstrainedToScreen: PropTypes.bool,
-  /** Title of thing that is being described. */
+  /** Id of the thing that is being described. */
+  popoverId: PropTypes.string,
+  /** Title of the thing that is being described. */
   popoverTitle: PropTypes.string,
   /** Test section */
   testSection: PropTypes.string,